Add a message used for discovering the local network interface address.
Devices need to announce their address, yet this is hard to determine
correctly without actively sending out a message and asking the
receiver what the sending address was.

This adds a protocol message "/address" which does exactly that -- any
hosts that receive /address with a single integer, (the random ID),
must respond with the ID and the sender's hostname.  The sender will
receive these responses and memorize its address.

The device is not registered until it knows its address.  Therefore if
there is no response, the device will continue to send out "/address"
requests every 2 seconds.

Pros:

- Correctly identify address according to how receivers see you.

Cons:

- Extra network traffic.

- What if a device on the same host responds first? There is no
  guarantee that the address it responds with will be useful to the
  rest of the LAN.

- Lone devices will never register.  Not necessarily a problem, since
  registration is only useful when someone else is around to notice.
  However, problematic for the tests and examples that (perhaps
  misleadingly) wait on mdev_ready() at startup.